Cambiar el codigo de error

+++ 2_Back_Python_FastAPI_Otro_Curso_8Hrs/Ejercicio_3_HTTP_Status_Code/Practica_1/model/celular.py
@@ -2,11 +2,11 @@ from pydantic import BaseModel
 
 class Celular(BaseModel):
     id : int
-    peso : int
+    peso : float
     color : str
     marca : str
     operador : str
-    dimesion_x : int
-    dimesion_y : int
+    dimesion_x : float
+    dimesion_y : float
